Small modification to algorithm
Takes care of variables appearing in the code definition of other
variables... ffs soon I'll have to write my own language parser lol, if
this keeps up. This is a quick fix, if variables with names that contain
the name of other variables are used, this could cause an issue when
replacing. (I'll add an issue to fix that in the future... either only
allow full variable name replacement, or more easily, start replacing
from longer to smaller variable names)